+++ generator/integration_tests/golden/golden_thing_admin_connection.h
@@ -21,7 +21,7 @@
 
 #include "generator/integration_tests/golden/golden_thing_admin_connection_idempotency_policy.h"
 #include "generator/integration_tests/golden/internal/golden_thing_admin_stub.h"
-#include "generator/integration_tests/golden/retry_traits.h"
+#include "generator/integration_tests/golden/internal/retry_traits.h"
 #include "google/cloud/backoff_policy.h"
 #include "google/cloud/future.h"
 #include "google/cloud/options.h"
