+++ generator/integration_tests/golden/google_cloud_cpp_generator_golden_files.bzl
@@ -51,8 +51,8 @@ google_cloud_cpp_generator_golden_files = [
     "internal/golden_thing_admin_stub.h",
     "internal/golden_thing_admin_stub_factory.cc",
     "internal/golden_thing_admin_stub_factory.h",
+    "internal/retry_traits.h",
     "mocks/mock_golden_kitchen_sink_connection.h",
     "mocks/mock_golden_thing_admin_connection.h",
-    "retry_traits.h",
     "streaming.cc",
 ]
