+++ google/cloud/bigquery/google_cloud_cpp_bigquery.bzl
@@ -27,6 +27,7 @@ google_cloud_cpp_bigquery_hdrs = [
     "internal/bigquery_read_option_defaults.h",
     "internal/bigquery_read_stub.h",
     "internal/bigquery_read_stub_factory.h",
+    "internal/retry_traits.h",
     "retry_traits.h",
 ]
