+++ google/cloud/pubsublite/google_cloud_cpp_pubsublite.bzl
@@ -27,7 +27,7 @@ google_cloud_cpp_pubsublite_hdrs = [
     "internal/admin_option_defaults.h",
     "internal/admin_stub.h",
     "internal/admin_stub_factory.h",
-    "retry_traits.h",
+    "internal/retry_traits.h",
 ]
 
 google_cloud_cpp_pubsublite_srcs = [
