+++ google/cloud/spanner/admin/database_admin_connection.h
@@ -21,7 +21,7 @@
 
 #include "google/cloud/spanner/admin/database_admin_connection_idempotency_policy.h"
 #include "google/cloud/spanner/admin/internal/database_admin_stub.h"
-#include "google/cloud/spanner/admin/retry_traits.h"
+#include "google/cloud/spanner/admin/internal/retry_traits.h"
 #include "google/cloud/backoff_policy.h"
 #include "google/cloud/future.h"
 #include "google/cloud/options.h"
